feat: add BaseButton component
